requirements.txt
torch==2.7.1
torchvision==0.22.1
ultralytics==8.3.163
opencv-python==4.11.0.86
pillow==11.2.1
clip @ git+https://github.com/ultralytics/CLIP.git@a6238961f9fab312d99c7709f8c7485b3f9a9b50
